fix: issue additional types are not recognized by worker
