+++ compass.yaml.example
@@ -52,8 +52,6 @@ service:
         max_send_msg_size: 33554432
         max_recv_msg_size: 33554432
         request_timeout: 5s
-    additional_asset_types:
-        - fact_source
 
 worker:
     enabled: true
@@ -75,3 +73,7 @@ client:
     host: localhost:8081
     serverheaderkey_uuid: Compass-User-UUID // if ommited, will use value on service.identity.headerkey_uuid
     serverheadervalue_uuid: gotocompany@email.com
+
+asset:
+    additional_types:
+        - fact_source
